# perf: Performance analysis and optimizations for Azure DevOps and self-maintenance