… docs/COMMAND_MAPPING.md
… docs/PERFORMANCE-ANALYSIS-REPORT.md
… docs/PROVIDER_STRATEGY.md
… docs/SELF-MAINTENANCE-GUIDE.md
… scripts/benchmarks/provider-router.bench.js
… scripts/benchmarks/self-maintenance-validate.bench.js
… test/regression/__snapshots__/health-report.json
… test/regression/__snapshots__/hybrid-strategy-latest.json
… wiki/CLI-Reference.md